Make SpawnedSigner generic so we can have either a v1 or v2 signer test easily spun up with duplicate code minimized
