+++ .github/workflows/bitcoin-tests.yml
@@ -82,15 +82,15 @@ jobs:
           - tests::nakamoto_integrations::vote_for_aggregate_key_burn_op
           - tests::nakamoto_integrations::follower_bootup
           - tests::nakamoto_integrations::forked_tenure_is_ignored
-          - tests::signer::stackerdb_dkg
-          - tests::signer::stackerdb_sign_request_rejected
-          - tests::signer::stackerdb_block_proposal
-          - tests::signer::stackerdb_filter_bad_transactions
+          - tests::signer::v1::dkg
+          - tests::signer::v1::sign_request_rejected
+          - tests::signer::v1::block_proposal
+          - tests::signer::v1::filter_bad_transactions
+          - tests::signer::v1::delayed_dkg
           # TODO: enable these once v1 signer is fixed
-          # - tests::signer::stackerdb_mine_2_nakamoto_reward_cycles
-          # - tests::signer::stackerdb_sign_after_signer_reboot
+          # - tests::signer::v1::mine_2_nakamoto_reward_cycles
+          # - tests::signer::v1::sign_after_signer_reboot
           - tests::nakamoto_integrations::stack_stx_burn_op_integration_test
-          - tests::signer::stackerdb_delayed_dkg
           # Do not run this one until we figure out why it fails in CI
           # - tests::neon_integrations::bitcoin_reorg_flap
           # - tests::neon_integrations::bitcoin_reorg_flap_with_follower
